+++ TPO/Api/src/main/java/com/uade/api/models/MovimientoReclamoModel.java
@@ -15,7 +15,7 @@ import lombok.*;
 public class MovimientoReclamoModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idMovimiento;
+    private Long idMovimiento;
     @ManyToOne
     @JoinColumn(name="Reclamo")
     private ReclamoModel reclamo;
